requirements.txt
# external requirements
flask
flask-cors
flask_httpauth
cianparser
scikit-learn
pandas
boto3
python-dotenv
PyGithub
jupyter
matplotlib
gunicorn
dvc[s3]
geopy
xgboost